paris: rename paris to paris_macros
This will distinguish between official crate and our module.
